Tested with real device
Module tested with real device.

Implemented HDMI switching for HDMI output.

Tested and validated with AMX NX-1200 + CYP PUV10824k22 matrix

Module is ready for deployment.
No feedback is provided by the machine, the module will report as feedback the last configuration.

No disconnect function found on the device, this means that you can only patch an input to an output and there is no possibility to disconnect the input from the outpu.